+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Changelog for the gruf gem. This includes internal history before the gem was ma
 
 ### Pending release
 
+### 2.4.0
+
+- Added a hash of error log levels to Interceptor class, mapping error code to level of logging to use.
+- To override the level of logging per error response, provide a map of codes to log level in options, key :log_levels. 
+- Default is :error log level.
+
 ### 2.3.0
 
 - Add Gruf::Interceptors::ClientInterceptor for intercepting outbound client calls